refactor: combine implementations of useCoState and useCoStates
